+++ module-api/src/main/java/com/wsws/moduleapi/chat/dto/ChatRoomApiResponse.java
@@ -1,6 +1,6 @@
-package com.wsws.moduleapi.dto.chat;
+package com.wsws.moduleapi.chat.dto;
 
-import com.wsws.moduleapplication.dto.chat.ChatRoomServiceResponse;
+import com.wsws.moduleapplication.chat.dto.ChatRoomServiceResponse;
 import com.wsws.moduleapplication.dto.user.UserProfileResponse;
 import com.wsws.moduledomain.user.vo.Nickname;
