+++ module-application/src/main/java/com/wsws/moduleapplication/chat/dto/ChatRoomServiceRequest.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.dto.chat;
+package com.wsws.moduleapplication.chat.dto;
 
 public record ChatRoomServiceRequest(String userId, String userId2) {
 }
